fix(games-create): silence remaining hydration warnings on /games/create (ARC-744)
Two distinct hydration mismatches were still firing on first paint:

- GlimwormPoster: the star field used Math.sin to seed positions, which
  is implementation-defined and drifts in the last digits between the
  Node V8 used for SSR and the Chrome V8 used at hydration. The server
  ended up rendering `cx="13.732360643916763"` while the client computed
  `cx={13.732360645662993}`. Round x/y/opacity to fixed precision so
  both sides serialize to identical strings.

- Rail Critical card cluster: the real `<HandCard>` fan rendered via
  Tamagui in SSR emits longhand `background-image` / `border-style`
  declarations while the client uses the `background` shorthand, which
  React flags as a tree mismatch. Extract the 5-card fan into
  CriticalCardClusterReal and dynamic-import it with `ssr: false`, the
  same pattern already used for SeaBattleRealPreview. The SVG poster
  continues to render during SSR + first paint; the real cards mount
  client-side only.
