+++ apps/web/src/features/games/ui/create/redesign/RailPreviewArt.tsx
@@ -1,11 +1,9 @@
 'use client';
 
-import { useMemo, type CSSProperties } from 'react';
+import { useMemo } from 'react';
 import dynamic from 'next/dynamic';
-import { HandCard } from '@/widgets/CriticalGame/ui/hand/HandCard';
 import { handWithUids } from '@/widgets/CriticalGame/lib/combo';
 import type { CriticalCard } from '@/widgets/CriticalGame/types';
-import s from './GameCreateView.module.css';
 import { GameArt } from './art/GameArt';
 import { CriticalCardPoster } from './art/CriticalCardPoster';
 import { SeaBattleBoardPoster } from './art/SeaBattleBoardPoster';
@@ -19,6 +17,11 @@ import { useSpriteLoaded } from './useSpriteLoaded';
 const SeaBattleRealPreview = dynamic(() => import('./SeaBattleRealPreview'), {
   ssr: false,
 });
+
+const CriticalCardClusterReal = dynamic(
+  () => import('./CriticalCardClusterReal'),
+  { ssr: false },
+);
 
 interface Props {
   gameId: GameId;
@@ -102,69 +105,8 @@ function CriticalCardCluster({ themeId }: { themeId: string }) {
           pointerEvents: showReal ? 'auto' : 'none',
         }}
       >
-        <RealCardCluster cards={cards} themeId={themeId} />
+        <CriticalCardClusterReal cards={cards} themeId={themeId} />
       </div>
-    </div>
-  );
-}
-
-interface ClusterCardsProps {
-  cards: ReturnType<typeof handWithUids>;
-  themeId: string;
-}
-
-// Five-card fan, left→right. Outer two are dimmed/blurred ghosts, inner two
-// are partially dimmed, center card is the hero.
-const FAN_SLOTS = [
-  { x: -116, y: 18, rotate: -18, opacity: 0.45, scale: 0.78, z: 0 },
-  { x: -62, y: 6, rotate: -10, opacity: 0.7, scale: 0.86, z: 1 },
-  { x: 0, y: 0, rotate: -1, opacity: 1, scale: 0.95, z: 2 },
-  { x: 62, y: 6, rotate: 10, opacity: 0.7, scale: 0.86, z: 1 },
-  { x: 116, y: 18, rotate: 18, opacity: 0.45, scale: 0.78, z: 0 },
-] as const;
-
-function RealCardCluster({ cards, themeId }: ClusterCardsProps) {
-  return (
-    <div
-      style={{
-        position: 'relative',
-        width: '100%',
-        height: '100%',
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'center',
-      }}
-    >
-      {cards.map((card, i) => {
-        const slot = FAN_SLOTS[i] ?? FAN_SLOTS[0];
-        const isHero = i === 2;
-        // Drive the per-card position/rotation through CSS variables so the
-        // `.fanCard:hover` rule in the module CSS can lift the hovered card
-        // (the same gesture as the home-hero card stack) without losing the
-        // card's resting rotation / scale.
-        const slotStyle: CSSProperties = {
-          ['--fan-x' as string]: `${slot.x}px`,
-          ['--fan-y' as string]: `${slot.y}px`,
-          ['--fan-rotate' as string]: `${slot.rotate}deg`,
-          ['--fan-scale' as string]: String(slot.scale),
-          ['--fan-opacity' as string]: String(slot.opacity),
-          ['--fan-z' as string]: String(slot.z),
-          ['--fan-filter' as string]: isHero ? 'none' : 'blur(0.4px)',
-        };
-        return (
-          <div key={card.uid} className={s.fanCard} style={slotStyle}>
-            <HandCard
-              card={card}
-              isSelected={false}
-              disabled
-              cardVariant={themeId}
-              showName
-              showDescription={false}
-              onToggle={() => {}}
-            />
-          </div>
-        );
-      })}
     </div>
   );
 }
